Changes architecture of researcher, aide, and hold list classes.
Changes the hold list class to use an array instead of a queue to avoid threading issues.  Changes the file loading in the aide spec.  Changes the researcher class to be responsible only for getting the next search term from the hold list and using it to perform a search.